Reorder sources
